reorganization: creation of Enzo/fluid-props
